pyproject.toml
[tool.poetry]
name = "ecommerce-marketing-assistant"
version = "0.1.0"
description = ""
authors = ["AhernShaia <gemgt4gtii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
azure-cognitiveservices-vision-computervision = "^0.9.0"
openai = "^1.14.3"
diffusers = "^0.27.2"
torch = "^2.3.0"
torchvision = "^0.18.0"
langchain-community = "^0.0.34"
line-bot-sdk = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
